test(ui): 패널 폭 검증 위치를 정리
